refactor: move TextBase to _private directory
