+++ src/components/text/_private/TextBase/TextBase.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/_partials/index';
+@import '../../../../styles/_partials/index';
 
 .TextBase {
 	@include defaultFontSettings();
