+++ src/components/text/_private/TextBase/TextBase.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import classnames from 'classnames';
 import * as styles from './TextBase.scss';
-import ILocalContainerProps from '../../../common/structures/ILocalContainerProps';
-import { Container } from '../../Container';
+import ILocalContainerProps from '../../../../common/structures/ILocalContainerProps';
+import { Container } from '../../../Container';
 
 export enum TextBasePropColor {
 	gray = 'gray',
