+++ src/components/text/_private/TextBase/examples/TextBaseExample.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import IReactComponentProps from '../../../../common/structures/IReactComponentProps';
-import { ComponentExampleBase } from '../../../../common/helpers/ComponentExampleBase';
+import IReactComponentProps from '../../../../../common/structures/IReactComponentProps';
+import { ComponentExampleBase } from '../../../../../common/helpers/ComponentExampleBase';
 import TextBase, { TextBasePropFontSize, TextBasePropFontWeight } from '../TextBase';
 
 export class TextBaseExample extends ComponentExampleBase {
